# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Will Be Done API",
    "description": "Public HTTP API for Will Be Done.",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token returned by the login or register endpoint."
      }
    },
...
  },
  "paths": {
...
    "/api/v1/spaces/{spaceId}/sections/{sectionId}/task-templates": {
      "post": {
        "operationId": "createTaskTemplate",
        "summary": "Create a task template",
        "tags": ["Task templates"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string", "minLength": 1 },
                  "content": {
                    "anyOf": [{ "type": "string" }, { "type": "null" }]
                  },
                  "nature": {
                    "anyOf": [
                      { "type": "string", "enum": ["red", "green", "unknown"] },
                      { "type": "null" }
                    ]
                  },
                  "repeatRule": { "type": "string", "minLength": 1 },
                  "repeatRuleDtStart": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9007199254740991
                  },
                  "placement": {
                    "oneOf": [
                      {
                        "type": "object",
                        "properties": {
                          "kind": { "type": "string", "enum": ["first"] }
                        },
                        "required": ["kind"],
                        "additionalProperties": false
                      },
                      {
                        "type": "object",
                        "properties": {
                          "kind": { "type": "string", "enum": ["last"] }
                        },
                        "required": ["kind"],
                        "additionalProperties": false
                      },
                      {
                        "type": "object",
                        "properties": {
                          "kind": { "type": "string", "enum": ["before"] },
                          "anchorId": {
                            "type": "string",
                            "minLength": 1,
                            "description": "Sibling to place the entity before"
                          }
                        },
                        "required": ["kind", "anchorId"],
                        "additionalProperties": false
                      },
                      {
                        "type": "object",
                        "properties": {
                          "kind": { "type": "string", "enum": ["after"] },
                          "anchorId": {
                            "type": "string",
                            "minLength": 1,
                            "description": "Sibling to place the entity after"
                          }
                        },
                        "required": ["kind", "anchorId"],
                        "additionalProperties": false
                      }
                    ],
                    "description": "Position within an ordered collection"
                  }
                },
                "required": ["title"],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": { "type": "string", "minLength": 1 },
            "in": "path",
            "name": "spaceId",
            "required": true,
            "description": "Space identifier"
          },
          {
            "schema": { "type": "string", "minLength": 1 },
            "in": "path",
            "name": "sectionId",
            "required": true,
            "description": "Project section identifier"
          }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "template": {
                      "type": "object",
                      "properties": {
                        "type": { "type": "string", "enum": ["template"] },
                        "id": {
                          "type": "string",
                          "description": "Task template identifier"
                        },
                        "title": { "type": "string" },
                        "content": { "type": "string" },
                        "projectSectionId": {
                          "type": "string",
                          "description": "Parent project section identifier"
                        },
                        "nature": {
                          "type": "string",
                          "enum": ["red", "green", "unknown"]
                        },
                        "repeatRule": { "type": "string" },
                        "repeatRuleDtStart": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 9007199254740991
                        },
                        "createdAt": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 9007199254740991
                        },
                        "lastGeneratedAt": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 9007199254740991
                        }
                      },
                      "required": [
                        "type",
                        "id",
                        "title",
                        "projectSectionId",
                        "nature",
                        "repeatRule",
                        "repeatRuleDtStart",
                        "createdAt",
                        "lastGeneratedAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": ["template"],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "BAD_REQUEST",
                        "UNAUTHORIZED",
                        "FORBIDDEN",
                        "NOT_FOUND",
                        "CONFLICT",
                        "INTERNAL_SERVER_ERROR"
                      ]
                    },
                    "message": { "type": "string" }
                  },
                  "required": ["code", "message"],
                  "additionalProperties": false,
                  "description": "Error response"
                }
              }
            }
          },
          "403": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "BAD_REQUEST",
                        "UNAUTHORIZED",
                        "FORBIDDEN",
                        "NOT_FOUND",
                        "CONFLICT",
                        "INTERNAL_SERVER_ERROR"
                      ]
                    },
                    "message": { "type": "string" }
                  },
                  "required": ["code", "message"],
                  "additionalProperties": false,
                  "description": "Error response"
                }
              }
            }
          },
          "404": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "BAD_REQUEST",
                        "UNAUTHORIZED",
                        "FORBIDDEN",
                        "NOT_FOUND",
                        "CONFLICT",
                        "INTERNAL_SERVER_ERROR"
                      ]
                    },
                    "message": { "type": "string" }
                  },
                  "required": ["code", "message"],
                  "additionalProperties": false,
                  "description": "Error response"
                }
              }
            }
          },
          "409": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "BAD_REQUEST",
                        "UNAUTHORIZED",
                        "FORBIDDEN",
                        "NOT_FOUND",
                        "CONFLICT",
                        "INTERNAL_SERVER_ERROR"
                      ]
                    },
                    "message": { "type": "string" }
                  },
                  "required": ["code", "message"],
                  "additionalProperties": false,
                  "description": "Error response"
                }
              }
            }
          },
          "500": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "BAD_REQUEST",
                        "UNAUTHORIZED",
                        "FORBIDDEN",
                        "NOT_FOUND",
                        "CONFLICT",
                        "INTERNAL_SERVER_ERROR"
                      ]
                    },
                    "message": { "type": "string" }
                  },
                  "required": ["code", "message"],
                  "additionalProperties": false,
                  "description": "Error response"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/spaces/{spaceId}/task-templates/{templateId}/move": {
      "post": {
        "operationId": "moveTaskTemplate",
        "summary": "Move a task template",
        "tags": ["Task templates"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "projectSectionId": { "type": "string", "minLength": 1 },
                  "placement": {
                    "oneOf": [
                      {
                        "type": "object",
                        "properties": {
                          "kind": { "type": "string", "enum": ["first"] }
                        },
                        "required": ["kind"],
                        "additionalProperties": false
                      },
                      {
                        "type": "object",
                        "properties": {
                          "kind": { "type": "string", "enum": ["last"] }
                        },
                        "required": ["kind"],
                        "additionalProperties": false
                      },
                      {
                        "type": "object",
                        "properties": {
                          "kind": { "type": "string", "enum": ["before"] },
                          "anchorId": {
                            "type": "string",
                            "minLength": 1,
                            "description": "Sibling to place the entity before"
                          }
                        },
                        "required": ["kind", "anchorId"],
                        "additionalProperties": false
                      },
                      {
                        "type": "object",
                        "properties": {
                          "kind": { "type": "string", "enum": ["after"] },
                          "anchorId": {
                            "type": "string",
                            "minLength": 1,
                            "description": "Sibling to place the entity after"
                          }
                        },
                        "required": ["kind", "anchorId"],
                        "additionalProperties": false
                      }
                    ],
                    "description": "Position within an ordered collection"
                  }
                },
                "required": ["projectSectionId", "placement"],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": { "type": "string", "minLength": 1 },
            "in": "path",
            "name": "spaceId",
            "required": true,
            "description": "Space identifier"
          },
          {
            "schema": { "type": "string", "minLength": 1 },
            "in": "path",
            "name": "templateId",
            "required": true,
            "description": "Task template identifier"
          }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "template": {
                      "type": "object",
                      "properties": {
                        "type": { "type": "string", "enum": ["template"] },
                        "id": {
                          "type": "string",
                          "description": "Task template identifier"
                        },
                        "title": { "type": "string" },
                        "content": { "type": "string" },
                        "projectSectionId": {
                          "type": "string",
                          "description": "Parent project section identifier"
                        },
                        "nature": {
                          "type": "string",
                          "enum": ["red", "green", "unknown"]
                        },
                        "repeatRule": { "type": "string" },
                        "repeatRuleDtStart": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 9007199254740991
                        },
                        "createdAt": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 9007199254740991
                        },
                        "lastGeneratedAt": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 9007199254740991
                        }
                      },
                      "required": [
                        "type",
                        "id",
                        "title",
                        "projectSectionId",
                        "nature",
                        "repeatRule",
                        "repeatRuleDtStart",
                        "createdAt",
                        "lastGeneratedAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": ["template"],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "BAD_REQUEST",
                        "UNAUTHORIZED",
                        "FORBIDDEN",
                        "NOT_FOUND",
                        "CONFLICT",
                        "INTERNAL_SERVER_ERROR"
                      ]
                    },
                    "message": { "type": "string" }
                  },
                  "required": ["code", "message"],
                  "additionalProperties": false,
                  "description": "Error response"
                }
              }
            }
          },
          "403": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "BAD_REQUEST",
                        "UNAUTHORIZED",
                        "FORBIDDEN",
                        "NOT_FOUND",
                        "CONFLICT",
                        "INTERNAL_SERVER_ERROR"
                      ]
                    },
                    "message": { "type": "string" }
                  },
                  "required": ["code", "message"],
                  "additionalProperties": false,
                  "description": "Error response"
                }
              }
            }
          },
          "404": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "BAD_REQUEST",
                        "UNAUTHORIZED",
                        "FORBIDDEN",
                        "NOT_FOUND",
                        "CONFLICT",
                        "INTERNAL_SERVER_ERROR"
                      ]
                    },
                    "message": { "type": "string" }
                  },
                  "required": ["code", "message"],
                  "additionalProperties": false,
                  "description": "Error response"
                }
              }
            }
          },
          "409": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "BAD_REQUEST",
                        "UNAUTHORIZED",
                        "FORBIDDEN",
                        "NOT_FOUND",
                        "CONFLICT",
                        "INTERNAL_SERVER_ERROR"
                      ]
                    },
                    "message": { "type": "string" }
                  },
                  "required": ["code", "message"],
                  "additionalProperties": false,
                  "description": "Error response"
                }
              }
            }
          },
          "500": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "BAD_REQUEST",
                        "UNAUTHORIZED",
                        "FORBIDDEN",
                        "NOT_FOUND",
                        "CONFLICT",
                        "INTERNAL_SERVER_ERROR"
                      ]
                    },
                    "message": { "type": "string" }
                  },
                  "required": ["code", "message"],
                  "additionalProperties": false,
                  "description": "Error response"
                }
              }
            }
          }
        }
      }
    },
...
  }
}